Enable `Children = [AnyNode]` for Button and Label components (#62)
When in the future we add more components that would be useful as children of `Button` and `Label`, for example components that render to `CAShapeLayer`, we won't be able to add them as children. The reason is that this node "slot" is taken by text of `String` type. I think it's better to make this API-breaking change before the initial release, even though we don't have a component for `CAShapeLayer` yet.

* Move `text` from Children to Props in Button/Label
* Fix unit-tests after text props change