Add factory functions creating `SizeConstraint` (#67)
We had an existing `SizeConstraint` struct, which was correctly rendered to auto layout constraints, but there was no way for users to create `SizeConstraint` with the new auto layout syntax. This is fixed by adding extension functions on `Size` with the usual format. This now allows setting constraints for size with a constant value:

```swift
Size.equal(to: .init(width: 100, height: 200))
```

Or with a multiplier relative to other nodes:

```swift
Size.equal(to: .parent, multiplier: 0.5)
```

* Add factory functions creating `SizeConstraint`
* Apply fixes from new SwiftFormat